refactor(publish): extract shared model getters from excel writer
Move cached getters, location suffix helpers, and cache clearing into
model_objects and cache_reset so XLSX and other publishers share one path.
