+++ bedrock/extract/disaggregation/export_cornerstone_waste_disagg_matrices.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import os
 import pathlib
 
-from bedrock.publish.excel.writer import clear_publish_caches
+from bedrock.publish.model_objects import clear_publish_caches
 from bedrock.transform.allocation.derived import derive_E_usa
 from bedrock.transform.eeio.cornerstone_disagg_pipeline import get_waste_disagg_weights
 from bedrock.transform.eeio.derived_cornerstone import (
